// client/app/shared/models/appplication-models/steps.model.ts
export class Step1 {
    firstName?: string = '';
    middleName?: string = '';
    lastName?: string = '';
    email?: string = '';
    dob?: string = '';
    currentAddress1?: string = '';
    currentAddress2?: string = '';
    city?: string = '';
    state?: string = '';
    zipCode?: string = '';
    primaryPhone?: string = '';
    primaryPhoneType?: string = '';
    alternatePhone?: string = '';
    alternatePhoneType?: string = '';
    residencyStatus?: string = '';
}

export class Step2 {
    propertyType?: string = '';
    propertyAttached?: string = '';
    useOfProperty?: string = '';
    purchasePrice?: string = '';
    downPayment?: string = '';
    annualTaxes?: string = '';
    annualHazard?: string = '';
    annualFlood?: string = '';
    agent?: string = '';
    homePlan?: string = '';
    purchaseAgreement?: string = '';
    equity?: string = '';
}

export class Step3{
    annualIncome?: string = '';
    totalAssets?: string = '';
}